Move announce logic into `TagListItemMixin`
